Made instance state values in the CSV summary translatable
The instance state values ("Active", "Stopped" and others)
captured in the CSV Summary from the Project Overview page
were not translatable.  Two changes were needed.  Some of the
possible values from Nova were missing from the STATUS_DISPLAY_CHOICES
tuple and needed to be added.  Then the state values
needed to be translated into the proper locale when the CSV was
built.  It should be noted that I used the same style Lucas Palm
proposed in change 226579 (launchpad 1490031) for the second change.

Change-Id: Ib52d4c1fade22199390c6a568fc8f1101784a88d
Closes-Bug: #1496953