Fix 3-components accessors for TabbedTables in i9n tests
When using just 2 last segments of the path to table for generating
accessor method name it's possible that resulting names will clash,
say for Project->Compute->Volumes->Volumes and for
Admin->System->Volumes->Volumes. To avoid this situation, we need to
use 3 components name for tables which are at 4th level of nesting.

Change-Id: If63ae14a99f4c53030927f96a69b7189bd83c4e5
Closes-Bug: #1467615