Set autodismissAlert on each alert
Alerts that are spawned from javascript aren't auto-dismissed because
horizon.autoDismissAlerts is only called once on page load (the network
topology JS calls it explictly to work around this). This patch sets the
autodismiss on each alert as it's created if applicable so that alerts
created by JS are also auto-dismissed, and calls the function on any
alerts created statically (i.e. in django code) at page load.

Change-Id: I7d42ba4fbc7565fb43fb0cc894154100e8fe0fe9
Closes-Bug: #1442288