Adding Magic Search codebase to Horizon
Magic Search is getting integrated into the Horizon codebase as agreed
at the Tokyo summit.  This Magic Search patch replaces the XStatic
package with its own module, and otherwise works as the current XStatic
package (changes will be made to accommodate Searchlight features in
following patches).  This is not simply a copy of what is in GitHub,
because there were no tests and little documentation.

The code underwent significant restructuring to accommodate the linter and
other standards used in Horizon.

Change-Id: I9a2b0f3fed1955680a534e8d284da2c8ee68ef16
Implements: blueprint integrate-magic-search