Basic Tabs now inherit from a Bootstrap Theme
Tabs were pretty basic.  For the most, tabs were using the standard
styles nicely, except when embedded inside of a modal. There was a
lot of adding padding and removing padding to the tab-content and
modal container.  This has been cleaned up.

Also, the 'asterisk' that is used to show when a tab has required
information for a workflow was changed to an icon font.

Many unused styles were removed from horizon.scss.

Partially-Implements: blueprint horizon-theme-css-reorg
Partially-Implements: blueprint bootstrap-html-standards

Change-Id: I929de0ae31a3c2cca9e332956eed755c4e279f3d