Remove data-type and sort from flavor column
Due to the flavor column in the instances table having
a data-type of size, the sortable plugin tries to sort
it by a size parser.
Removing the sorter for that column seems the
sanest decision as the end-user may expect the
sorting to be done on cpu or disk or any number
of factors that we cannot know in advance.

Change-Id: I418244069bdcbaa2248d99772090806825987d39
Closes-Bug: 1518893