Networks tab does not show external configuration
When you create a network you can specify it as
external network. When the network is listed
in Horizon 'Networks' tab, we do not see any
column indicating if network is 'external' .
It is nice to enhance the network tab to add a
new column showing 'external' indicating whether
network is 'external'.

Change-Id: I6d07ffc0d2584bbf8bd6c25aee660f87c385e737
Closes-Bug:#1350237