ICMP type & code validation while adding Security Group rules
While adding new 'Security Group Rule' viz. 'Custom ICMP rule' when
wrong ICMP 'type' or 'code' was given the errors it would show were
'Not a valid port number'; which is misleading. The errors should
rather be 'Not a valid ICMP type' or 'Not a valid ICMP code'.

Also for validating ICMP 'type' and 'code' there wasn't any dedicated
functionality in 'oslo.utils/netutils' so the code for validating 'TCP
ports' was used. TCP ports are in range 0 to 65535 while ICMP 'type'
and 'code' are in range 0 to 255, so using 'TCP port validation' code
is incorrect.

When -1 is used in 'ICMP code' or 'ICMP type' that means any number
0 to 255 is valid.

Here newer dedicated functionality of 'oslo.utils/netutils' is used
for validating ICMP 'type' and ICMP 'code'.

Change-Id: I8e227a0021d418294fa7b7756d58e39f2100850a
Closes-Bug: #1511748