Do not use empty description fields
When editing the user by the user Update Form
make sure that if the description field has not changed
to not send it, as it will be saved as empty string.
Instead, the value should be absent from the user object,
so it respects the same values as users created in keystone
directly.
Unfortunately, inline editing still triggers the issue
as there is no good way of checking the initial value
without making an extra call.

Change-Id: Id2e0b3c589855068cbc15adc75b57ff0bf61a1b9
Partial-Bug: 1519579