hzIfApiVersion directive to check if api version enabled
Similar to the directives inside
openstack_dashboard/static/app/core/cloud-services, we want to remove
particular HTML content if not supported by the api version.

For example, we only want to show the users magic search if v3 is
enabled because only v3 supports the filtering.  Or we want to show
the Domain ID/Name on the Create Users modal if v3 is enabled.

You can do simple comparison checks, e.g. keystone version > 2
or glance <= 2 and chain together multiple checks.

Change-Id: Ided631a38d61347bb766fb11234917a29d2b5fe8
Closes-Bug: #1522520