Actions directive should take Service
Currently the actions directive uses callbacks and
permissions to determine what actions are allowed
and what callback to bind them.

The callback is passed as strings which result in
leaky abstractions and implicit binding of the
controller and the service with the HTML.

The patch changes the directive to only require a service
which needs to provide functions that help the directive
determine if the action is permitted and what method
to call when the action is performed.

Change-Id: I91f009c10c5db8bf0b0082625b259bf1ba7358c3
Partially-Implements: blueprint angularize-images-table