Remove keypair ID from details table
Currently Horizon shows keypdair.id in
project->compute->access&security->keypairs
-> keypair details tab.
But python-novaclient always returns
keypair.id=keypair.name.

Right now Nova sends ID in its response and python-novaclient got it,
but python-novaclient does not pass that ID into the Horizon.
And keypair-id is also shown via nova-cli.
But Nova folks says that ID is a some kind of private atrtr in Nova,
and that fact it is shown via cli is a side-effect and should be fiexed.

So, Horizon should not use keypair-id and this patch removes it.

Change-Id: Idd13b8900342fb98675fc8c98fff016109d8161b
Closes-Bug: #1527257