Clean up horizon.scss - Keypairs
A major part of the css reorg is a better organization of the css
styles themselves. horizon.scss should be a global import level file.

Keypair details was, for some reason, using a textarea element to
display non-form 'Public Key'.  Textarea is used for allowing users
to input data, this is a read only display. Replaced with a simple
<div> and removed the unnecessary styles from horizon.scss.  Added
a util class for easier use.

Cleaned up class name for 'key-text' class name.  Adding classes,
even if they are unused, allows for rich customization hooks for
consumers of Horizon.

Change-Id: I3d5ce1cb0a4cae6774e515206e511f61e2a9d10e
Partially-Implements: blueprint horizon-theme-css-reorg