Clean up horizon.scss - default/bootstrap/_type.scss
A major part of the css reorg is a better organization of the css
styles themselves. horizon.scss should be a global import level file.

page-header is a Bootstrap style.  We shouldn't be overriding it so
much.  Inherit from the theme.  Also, fixed padding of header on
default theme.

Change-Id: I771e03b3fc24ccd716bb556ecb8342e0da1dc37b
Partially-Implements: blueprint horizon-theme-css-reorg