Eliminate ambiguity when matching panel in i9n tests
To do so, an argument `src_elem` is added to _click_menu_item()
method. Whenever dashboard or a panel group is clicked, its wrapper is
returned to be used as `src_elem` in a subsequent call for clicking
third-level item.  This way the set of panel labels being matched is
restricted to the descendants of that particular dashboard or panel
group.

Change-Id: I54f3febed645b6bf2faddfbe27690ceb0944cd12
Closes-Bug: #1533160