Add Glance v1 / v2 capable angular image visibility filter
This patch provides a filter for evaluating the image visibility filter
which is compatible with both v1 (is_public) and v2 (visiblity) Glance API
results.

It also handles the ability to return a derived visibility of "Shared with Me"
based on if a passed in current project ID does not matches the image owner when
the image is set to private. In a discussion with a Glance core, this was said to
still be the correct way to accomplish determining if an image is shared or not.

Note: Although additional visibility statuses of shared and community
are proposed for Glance, they have not been agreed to by the Glance team
yet and it doesn't look like they'll make it anytime soon. Even when they do
make it, the logic in this filter will apply. It will just need to be updated
to accomodate whatever new algorithm the Glance service supports.

The filter is written to be as safe as possible to ensure
user friendly output while being as accurate as possible.

To test set DISABLED = False in _1051_project_ng_images_panel.py

Change-Id: If2b6ca72216d57543bb86fe91d7a6c8cf4667e6b
Partially-Implements: blueprint angularize-images-table