Fixed a few missed translation strings on Theme preview page
Under the Developer Dashboard, the Bootstrap Theme Preview has a few
strings that are not being translated when they could and should be.
Many of the strings are missing the "translate" directive to enable the
translation, while others include the directive but are not being translated
because of other issues.  The translation directive seems not to work when
there is another HTML element tag nested inside the outer element containing
the translation.  As such is the case with the Dropdown elements.  The
solution to this is to use the translate filter in situations where the
directive will not work.

This change fixes most of the translation issues on this page.

Note: The pseudo translation tool was used to determine what strings are
correctly being translated or not.

Change-Id: I5aeb9f6212f5ba6a1f4cde0b72cce638937b7138
Closes-Bug: #1536785