Angular Table Directive
Provide another way for developers to programatically declare
an Angular Table vs writing HTML markup.

You need to pass in the data, column definition, and item/row actions
into hz-dynamic-table.

The column definition includes
- selectAll
- expand
- a list of all the columns we want to display, For each column, provide
 name (to show data), title (what the column header is actually called),
priority, sortDefault, any filters on the column,
and an optional template. (Template is HTML markup for cells in that
specific column. It is used to show the data in a more complex manner than
just the name - item[column.name])

To test, pull down the dependent demo patch on the Users Table or the
Images table and set DISABLED = False in _3031_identity_users_panel.py
or _1051_project_ng_images_panel.py, respectively.
Then go to last Users panel in the Identity menu or the last Images table
in the Projects menu.

Change-Id: Id8ca0d13ed79030a85aaf3daaee1335ead5344f5
Partially-Implements: blueprint angular-table-directive