Updating JS unit test coverage threshold
This patch updates the JS unit test coverage threshold.  As we've added
code we've increased the coverage.  To keep new code well-covered, we
should increase the threshold so new code doesn't decrease the overall
coverage.  The rule of thumb is to keep it at the highest whole percentage
point that greater than .5% of a difference from the current coverage.

Change-Id: I89e18841b4e9ba4a2cbd49ee222d3234e9d26c88