Add allowed address pair extension UI for neutron ports.
Changed the port detail view to a TabbedTableView where the extra tab is
enabled/disabled when the extension is active or not in neutron. Similar to
how extensions are handled for routers. If the extension is not active the
port detail screen should look the same as it does now.

The extra tab has a table of the allowed address pairs (columns: IP address,
MAC) with create and delete actions.

Change-Id: I07edb1afae5c2004761d1c118a724fb94aaebe3e
implements: blueprint port-allowed-address-pairs-extension