Adding resource type registry service
In the Angular code we have established Actions that we are using in
table rows and headings.  These actions right now are collected by
individual custom services which are in turn used by pages that need
the actions (Image table page uses Image row and batch actions).

This patch eliminates the need to write custom services and instead
centralizes the registration of these actions.  On registraton they are
associated with a string which is a resource-type specification (used
in Searchlight, for example: 'OS::Glance::Image').  This makes the actions
available to any party needing to use them.

The actions are registered in modules' run clauses so they are
automatically registered, with the desired actions injected into the
registration.  This makes it easy for people to add their own actions
or their own resource types.

Change-Id: Ied6040b41513e9eaccecf59e82da1634df6ac08b
Partially-Implements: blueprint angularize-images-table