Add isCurrentProject to userSession service
Several times we've wanted to make a decision based on whether the current
user's project matches that of some given project on an item.  Rather than
rewriting this code all over the place, this patch centralizes the logic
so we can use this determination of ownership for all sorts of things
needing that kind of promise.

Change-Id: I281da92cfa5df7e04755097f81ede2b1b7ac835b
Partially-Implements: blueprint angularize-image-table