Adding Resource Type registration
This patch adds to the registry service to allow for resource types
to have their properties registered with configuration information.

It provides various convenience methods for outputting the property
labels and values, all determined through the registry, and provides
for singular and plural names to be defined.

This patch also registers many of the OpenStack resource types so
they may be used by implementations.  Ideally those are broken into
component modules, but until those are written they can reside here.

Change-Id: I71bebe253452944ff518f85b8840cfe921602544
Partially-Implements: blueprint angularize-images-table