Fixed over-application of webroot in disabled image panel
In haste to accommodate webroot settings for angular routing in the
images panel, we neglected to notice that the angular routing already
takes the base tag into account, and removes the base from the route path.
As a result, we can code a static path in and it will survive webroot
settings just fine.

To test: try using the Images panel with webroot both as / and as something
else (on your devstack).

Fixes-Bug: 1549530
Change-Id: I48479e88acf35c259016a615221c805bd46af291