Angular Policy Caching
With angular routing, every time a view / page is shown, any
policy checks are performed again. Policy decisions are fairly
static and should not change as long as the inputs are the same.

This patch makes it so that the results are cached until
a django page reload is triggered (such as when region, user,
project, panel are changed).

Change-Id: Ib8c133f680a7bf0357c26263fb26bb76eb13d65a
Closes-Bug: 1551815
Co-Authored-By: Matt Borland <matt.borland@hpe.com>