WIP: Fix property safety on magic search controller
It seems there are a few cases where properties are assumed to be on
the scope which are not present; particularly $scope.strings.  This
may be causing trouble with some of the features depending on how
they are initialized.

WIP: It's unknown if this fixes the problem as I cannot replicate it.

Change-Id: Icac2e032026d2e14939e9385f90bc34c18920fe1
Closes-Bug: 1552895