Images Actions Return Promises
The Angular images actions now consistently return a promise. This allows
a view to fetch an image related action from the registry, call it, and
have a consistent way to know that it is complete.

For actions, promises are preferred over events because they require
less setup and teardown, better handle the asynchronous nature of
modals, don't require a parent-child scope relationship and finally
allow for success/fail handler chaining.

Future patches may add data to the promise so that the view can know
which items were affected by the action.

This patch simply returns promises from actions as this is
a pattern I'd like to establish for Mitaka. Usage of these promises is
left to a follow-on patch. Also, removing the event mechanism from actions
is left to a follow-on patch.

Change-Id: I78d189d6eb026bd98abba6451641010104fc5002
Partially-Implements: blueprint angularize-images-table
Closes-Bug: 1555929
Partial-Bug: 1514627