Magic search shouldn't re-appear after selecting facet
When you select a facet, it instantly pops open again, obscuring the
results of the most recent facet.  This is fixed.

To test: Go to Launch Instance / Source, then click in the Available
search bar.  Add facets, and verify that after selecting each facet
opton that the facet drop-down is not shown.

Closes-bug: #1552594

Change-Id: Iffed933131163b6e64297ad8d7cd8912311c3ff0