Bypass get_domain call to keystone api
Warning messages in horizon and keystone logs regarding the 'Federated' domain
indicate that a call to get_domain is failing.

This fix will allow horizon to not attempt calls to retrieve domains
that do not exists.  The 'Federated' domain is a virtual domain that has
no record in the keystone database.

Change-Id: Ic3225815d12472d37c4105b656c5bc75b529c359
Closes-Bug: #1557791