Fix bad/inefficient API call for JS getImage()
The Angular API library for Glance has a function getImage().  This
function currently calls to an API like:

/api/glance/images/abasdfasdfas

However, that returns a 301 and redirects to:

/api/glance/images/abasdfasdfas/ <-- trailing slash

This causes latency and unnecessary overhead.

To test this fix, first verify the 301 problem before applying:

Run Horizon in debug (watch the server responses)
Enable the NG Images panel (enable _1051_ng_image...).
Go to the NG Images panel, then as you click on a Details view,
withess the 301 call with no trailing slash, followed by a
successful call with a 200 and trailing slash.

Then apply the patch.

Then do the same, but notice how there is only one 200 call, with
the trailing slash.

Change-Id: I0f119e2568d41c4aa63176a8bf8f2fabfbed1b47
Closes-Bug: 1559279