Disable all actions while performing action
This updates the angular actions controller so that all actions
controlled by the same actions controller become disabled immediately
when performing any action. This prevents the user from being able to
click an action a second time while the action is being initialized.
This is mainly a problem in cases where there is some asynchronous
operation happening that needs to resolve before the action can be
performed. For example, the Update Metadata action needs two API
calls to resolve before the modal is displayed. If the requests take
a long time to resolve or the user is feeling anxious they can click
the action a second time, or even some other action before the Update
Metadata modal even gets displayed.

This is not a perfect fix since there are a number of other bad
things a user could do if an action takes a long time such as
clicking a link or breadcrumb to navigate to another page. Ideally
actions should result in immediate feedback to the user and any
asynchronous operations happen afterwards. For example, the metadata
modal gets displayed immediately with a loading indicator and then
the API requests are initiated.

Closes-Bug: #1560303
Change-Id: I7d7696a03de85434f692ab0313d3dbdcbb582f25