Fixing updating user with no default project
The selector for setting the default project passes on an empty string
if the value is not changed. This is causing an error from keystone as
an empty string is not a valid project ID.

To fix, a couple of things have been done:

First, as user_update is a PATCH, we no longer pass on the default
project as a parameter in the update when using keystone v3
and the value did not change when editing the user.

Second, there is currently no way to unset the default project
using keystoneclient. An entry in the default project select
"select a project" has been removed when the default project is
already set. The entry was selectable and resulted in the
problematic empty string. The current default project is already
selected at load time. The ability to select a different default
project is maintained.

Additionally, keystone v2.0 functionality is not effected, as the
default project id is required in v2.0.

A test was added to cover the case of editing the default project.

Closes-Bug: #1561612
Change-Id: I3c84ee5c7af846b5a58f2dd0342a537d312323cf
(cherry picked from commit 32510d8211a43d83ebac6478fce17a8856cf9a0c)