Improve error message of when Conflict error occur in ng-swift
When we create a container, folder or object, error will occur if the
specified name is already exist. And when we delete, if specified
container or folder have a content, error will occur.
At the moment, error message is always same in all cases.
This patch will improve the message of when Conflict exception occur.

Change-Id: I825ef09badd1b10bb6fdab8f223bd6dfed28f3a4
Closes-bug: #1566820