Added a warning message if User Name is empty
Adding required=False to the user name form allows it to call
update_cell when it has been left empty.  I added logic to update_cell
to check if the form is empty or contains only whitespace and issue a
warning if true and only if the form in question is called 'name'.

Change-Id: Ieccb3f6d7be1acb4b7734e68d467a40cfc7eb99c
Closes-Bug: 1567398