Fix tests broken by novaclient 4.0.0 release
The release of python-novaclient 4.0.0 removed both volume_types
and volume_snapshots from the v2 namespace. Both of these data types
are supplied by cinder, but the Horizon test data was still populating
them in nova test data and the tests were using the nova values.

This patch removes the missing namespace references and replaces
the mocked test data with those mocked for cinder.

Additionally, the change exposed a rather broken test in the QoS
for volume_types test suite. This test has been fixed and improved
to exercise more branches.

This fix will unblock the gate for Horizon.

Closes-Bug: #1572246
Change-Id: I9604cd4046a5ab1486373d0572ddf504d4691196