Fix project name refresh in project menu bar
When makes an inline update to a project name the name in the menu bar
does not get updated after reloading/refreshing the page. This happens
because the menu display is retrieved from a token in the session which
does not get updated until a switch to a different project is made.

This fix addresses this issue by retrieving project name from the request
context variable authorized_tenants which has updated information instead
of session.

Fix doc string to confirm to pep8. Update template tests to account for
tenant name in menu bar. Tests are not setting session variable for tenant
name so previosuly template did not render active tenant name, this fix
does not use session anymore so tests need to account for it.

Change-Id: Ia0d5b51699197904d1a578d188762ef39a4b67cf
Closes-Bug: 1519577