table-detail (Angular) tables correct problem with row striping
A recent change introduced a problem with row striping for the tables
using table-detail.  This can be seen by going to the Launch Instance
(Angular) and looking at the different tables (Source, Flavor, Networks).
It's also well-shown when going to the Angular Images panel.

I corrected the problem seemingly by adding an explict default transparency
to all tr elements within this context.  This I believe overrides the
default bootstrap 2n selector, which would in this case gray out all odd lines,
which in this case would be all the primary (non-detail) rows.

I encountered another bug with the transfer table when you select the first
element, but that is separate from this issue and I'll file a bug on that.

Change-Id: Ic885df8c16ecb715fb7c8bc5e19bb1eead49a219
Closes-Bug: 1582732