Add *args, **kwargs to ModalBackdropMixin's init method
Before this change ModalBackdropMixin called super(...).__init__()
without arguments. This imposed restrictions on what classes this
mixin could have been mixed into, i.e. only classes, that do not accept
any parameters.
This proved to be a problem for murano-dashboard, since it uses this
mixin (indirectly through ModalFormMixin) and mixes it into a class,
that accepts parameters to it's init method.
This change allows to use ModalBackdropMixin with classes that have
init methods with parameters.

Change-Id: I6155476738021b784ef7e643c968f1d784b15906
Closes-Bug: #1582816