Set some useful default values with the new launch wizard.
In Mitaka, the new launch instance wizard regresses functionality
from the old launch wizard with regards to defaults. If there is
only one neutron network, it is not automatically selected, if
there is only one keypair, its not selected, and the default
security group is not automatically selected.

This patch fixes all of that.

Change-Id: I23bab6998e38ba1066b926a4fe71713768d9dff4
Closes-Bug: #1587681