In integration tests prevent getting volume size too early
In the test_volume_extend it was possible for a new volume size to be
fetched while the volume itself was still being extended, giving an
old size and failing a test as a result. Fix this intermittent failure
by waiting till the volume becomes 'Available' (and no longer
'Extending') before fetching its new size.

Closes-Bug: #1588307
Change-Id: I7197904dccd842eb4ef8931208b50ba2144c1a8c