The neutron policy file is out of date. This patch updates it to match neutron master.
Since the neutron policy was last updated, LBaaS, VPNaaS, and FWaaS,
have all been moved out of the neutron repo. When that was done,
apparently all policy support was removed as well. This patch retains
the related policy checks matching the old policy file rules. If
operators use the new policy file, the policy checks are harmless, as
the definition won't be found which will result in policy.check
returning True.

Additionally, the get_network call for the update network view was
modified to not have the subnet info populated as it's not used in
the form.

Change-Id: I6c40b99e88937d428a8e21fa28cdbc8a4190eb57