Migrate swift ui to use hz-dynamic-table
This removes the custom table code and uses the new
hz-dynamic-table directive to manage the table.

The documentation for actionResultService was also
edited to improve clarity.

Note: I was not intending to migrate all actions over
to use actionResultService in this patch (to keep the
patch size under control) so only the delete actions
have been done, and even then not optimally. These will
be revisited in a subsequent patch.

Change-Id: If8c4009c29fbbdbbeac12ce2bee4dcbef287ea98
Closes-Bug: 1576418
Closes-Bug: 1586175
Partially-Implements: blueprint swift-ui-functionality