Change default timeouts in integration tests
First, increase page_timeout to one minute, which should help in case
of a first couple of tests waiting for Apache workers to "warm up".

Second, reduce explicit_wait to 90 seconds (from 5 minutes), because
if something in Horizon doesn't complete in 90 seconds, then almost
for sure it won't complete in 5 minutes as well and only will spend
time, which is limited for integration tests.

Change-Id: Icb0767146fdb2682dca237ee18bbc4c3c8d976d0
Closes-Bug: #1592349