i18n: trim whitespaces in extracted messages from AngularJS templates
When extracting translation messages from AngularJS templates,
there is no need to keep whitespaces in the messages as there is
no meaning of repeated whitespaces in HTML.
This will make translation effort much simpler.
More detail is described in the bug report.

This commit trims such whitespaces. Django provides a convenient method
to do the same purpose for 'trimmed' option in Django templates.
This method is reused in this commit as well.

Closes-Bug: #1592965
Change-Id: I9b7ce54452f3db2350eecc3115db2e4173df5167