make breadcrumbs more robust
New swift UI does not define a 'page_title', but
template tags rely on it. The page will not show, if
not defined. This patch fixes it.

Note, in N release, this has significantly changed and a backport
of that new feature is not feasible. The issue itself is in
Mitaka only.

The change in N was
https://github.com/openstack/horizon/commit/d5b24d7b9760a68072cfed390c66a9433b415275

Change-Id: If56ae015d6795feb85094d1c701258966ecb9472