Adding hz-resource-property/-list and filter/value retrieval
This patch adds hz-resource-property, which outputs the given object
property as a label and value in a dt/dd.  This uses hz-cell which
ensures that all display rules dictated within the property
configuration are followed.

hz-resource-property-list is a level higher than hz-resource-property,
which allows for supplying multiple properties to be displayed, to
reduce excess markup.  It is not to be used at the exclusion of
hz-resource-property, but just as a way to make it easier to build
simple displays.

Change-Id: I62e3d9e4217aae428277bbef669ba704600f0b43
Partially-Implements: blueprint angular-registry