Show wait spinner after submit of wizard modal
Upon submitting a wizard modal, such as Launch Instance,
the modal remains open and the submit button is disabled.
This patch puts a spinner in the submit button as well to
show that work is going on in the background.

Note: Jenkins keeps failing due to firefox issues, so I'm
trying a different download link to see if it helps.

Change-Id: Ie800d9cc635d6b92af600e0bfe6846c70b6ca065
Closes-Bug: #1461641