Documentation and other improvements for registry
This patch improves the documentation of the
angular type registry.

It also removes use of "this" throughout the code
to reduce ambiguity, and moves the initAction method
to a more sensible place. The format() method is
no longer used (hz-cell formatting is used instead).

Finally it also adds an error log if the load function
isn't provided, otherwise developers could potentially
be chasing their tails by mis-configuring that feature.

Change-Id: I1b80971a4b19f26455dfc568de6bbb36b1245ef3
Partially-Implements: blueprint angular-registry