Support project column in admin view of NG images
This adds the project name column to ng images, which
is required in order to support an admin view for images.

In addition the hz-property directive could not handle
promises, which is required for doing things like
looking up the project name from the project id.
This adds limited support for that.

Change-Id: I0bb3782a23cf6ddff249459089b96a532aa8a0d8
Partially-Implements: blueprint angularize-images-table