Revert "Prevent creation of subnet via RBAC during new network creation"
The policy rule to create subnets requires a target to make the policy check. This implementation does not provide a target, so is not a proper place to use the workflow policy checks. To do this, a more involved approach needs to be used that supplies a target. However, this will be challenging because the target is the network_id that has yet to be created.

This reverts commit 3befade1411783af088b0e72b08f8ff3701b02c8.

Change-Id: I6af1cbd559034bd5635dfb4360b525714225677d