Remove array metadata when no items + fix case
Array metadata appends an operator automatically, which
means the value will still contain something and it
doesn't get detected for removal.

In addition, bad APIs like glance v1 will store a property
with mixed case, but then return only lower case when
listing them. However, it will not remove the property
when you request to remove it unless it matches the original
case (sad panda).  So, this patch will reset the proper
casing for properties defined as metadata definitions before
making the request to save the metadata.

Change-Id: I1e5a6fde35b7be6118f21ac46ad0aea088280215
Closes-Bug: 1606988